requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
kafka-python==2.0.2
redis==5.0.0
pybreaker==1.4.0
python-dotenv==1.0.0
numpy==1.24.3
